backend/requirements.txt
# Ghost-UI Backend Dependencies
# Versions locked for reproducible builds

# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
starlette==0.38.6

# Database
supabase==2.10.0

# AI Services
openai==1.54.0
google-generativeai==0.8.3

# Utilities
python-dotenv==1.0.1
pydantic[email]==2.9.2
email-validator==2.2.0
httpx==0.27.2

# Development
python-multipart==0.0.12

# Screenshots (for compliance scanner)
playwright==1.48.0

# Report generation
weasyprint==60.1
jinja2==3.1.2
